--- src/imwl/CMakeLists.txt
@@ -1,36 +1,0 @@
-add_executable(imwl
-    main.cpp
-    common.cpp common.h
-    Server.cpp Server.h
-    Compositor.cpp Compositor.h
-    Surface.cpp Surface.h
-    Seat.cpp Seat.h
-    Keyboard.cpp Keyboard.h
-    TextInputManagerV3.cpp TextInputManagerV3.h
-    TextInputV3.cpp TextInputV3.h
-    InputMethodManagerV2.cpp InputMethodManagerV2.h
-    InputMethodV2.cpp InputMethodV2.h
-    InputMethodKeyboardGrabV2.cpp InputMethodKeyboardGrabV2.h
-    InputPopupSurfaceV2.cpp InputPopupSurfaceV2.h
-    VirtualKeyboardManagerV1.cpp VirtualKeyboardManagerV1.h
-    VirtualKeyboardV1.cpp VirtualKeyboardV1.h
-    Xcb.cpp Xcb.h
-    X11KeyboardGrabber.cpp X11KeyboardGrabber.h
-    X11ActiveWindowMonitor.cpp X11ActiveWindowMonitor.h
-    DimTextInputManagerV1.cpp DimTextInputManagerV1.h
-    DimTextInputV1.cpp DimTextInputV1.h
-)
-
-target_link_libraries(imwl PUBLIC
-    Qt::Core
-    wls
-    PkgConfig::xcb
-    PkgConfig::xcb-xinput
-    PkgConfig::xkbcommon
-    shm_open_anon
-    $<$<BOOL:${Dtk6Core_FOUND}>:Dtk6::Core>
-)
-
-install(TARGETS imwl
-    BUNDLE BUNDLE DESTINATION .
-)
