--- src/imwl/common.h
@@ -1,12 +1,0 @@
-// SPDX-FileCopyrightText: 2023 UnionTech Software Technology Co., Ltd.
-//
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-#ifndef IMWL_COMMON_H
-#define IMWL_COMMON_H
-
-#include <stdint.h>
-
-uint32_t nextSerial();
-
-#endif // !IMWL_COMMON_H
